+ Type cast fix

+++ kex2/turok/renderer/renderSystem.cpp
@@ -321,7 +321,7 @@ void kexRenderSystem::SetOrtho(void) {
     dglMatrixMode(GL_MODELVIEW);
     dglLoadIdentity();
 
-    mtx.SetOrtho(0, sysMain.VideoWidth(), sysMain.VideoHeight(), 0, -1, 1);
+    mtx.SetOrtho(0, (float)sysMain.VideoWidth(), (float)sysMain.VideoHeight(), 0, -1, 1);
     dglLoadMatrixf(mtx.ToFloatPtr());
 
     dglMatrixMode(GL_PROJECTION);
